Move the exiftool to be per-lfr.
The single pattern-based invocation it previously had been using was
failing if the user provided .lfr files in different directories. This
new way will fix that, but also will increase the runtime overhead since
exiftool has to be initialized for every .lfr. This can be avoided by
leveraging the -stay_open flag [1], but my attempts at the use of it
proved to be more complex than I wanted to support.

[1]: http://www.sno.phy.queensu.ca/~phil/exiftool/exiftool_pod.html